requirements.txt
sentence-transformers
gradio
chromadb
langchain
langchain_core
langchain_community
einops
more_itertools
python-dotenv
langchain-google-genai